requirements.txt
Flask==3.0.3
pandas==2.2.2
google-generativeai==0.8.3
python-dotenv==1.0.1